backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.2
anthropic==0.75.0
python-dotenv==1.2.1